// MeskChatClient/src/types/responseEntityOfEmptyResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * MeskChatApplication.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */
import type { EmptyResponse } from './emptyResponse';
import type { ResponseEntityOfEmptyResponseValidationErrors } from './responseEntityOfEmptyResponseValidationErrors';

export interface ResponseEntityOfEmptyResponse {
  isSuccess?: boolean;
  /** @nullable */
  message?: string | null;
  data?: EmptyResponse;
  statusCode?: number;
  /** @nullable */
  validationErrors?: ResponseEntityOfEmptyResponseValidationErrors;
}